Utk data format loader.

+++ utkmap/src/interfaces.ts
@@ -35,7 +35,7 @@ export interface ILayerData {
 export interface ILayerGeometry {
     position: Float32Array;  // coordinate of the points
     normal?: Float32Array;   // normals of the points
-    indices?: Uint16Array;   // ids of the vertices
+    indices?: Uint32Array;   // ids of the vertices
 }
 
 export interface ILayerThematic {
